Added dependency injection for student model
