+++ MusicMaker/MusicMaker/View Controllers/ContainerViewController.swift
@@ -7,6 +7,8 @@
 //
 
 import UIKit
+import FirebaseFirestore
+import Firebase
 
 class ContainerViewController: UIViewController {
 
@@ -14,14 +16,17 @@ class ContainerViewController: UIViewController {
     // MARK: - Properties
     var teachersViewController: TeachersViewController!
     var sideMenuViewController: SideMenuViewController!
-
+    var student: Student?
+    let database = Firestore.firestore()
     
     // MARK: - IBOutlets
     @IBOutlet weak var sideMenu: UIView!
     @IBOutlet weak var teachersView: UIView!
     
+    // MARK: - View Life Cycle
     override func viewDidLoad() {
         super.viewDidLoad()
+        fetchStudent()
         self.sideMenu.transform = CGAffineTransform(translationX: -self.sideMenu.frame.width, y: 0)
         self.sideMenu.layer.shadowOpacity = 0.8
         sideMenuViewController = self.children[1] as? SideMenuViewController
@@ -30,10 +35,12 @@ class ContainerViewController: UIViewController {
         sideMenuViewController.delegate = self
     }
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        print(student?.email)
     }
     
+    // MARK: - Private Methods
     private func hideSideMenu() {
         UIView.animate(withDuration: 0.4, delay: 0, options: [], animations: {
             self.sideMenu.transform = CGAffineTransform(translationX: -self.sideMenu.frame.width, y: 0)
@@ -50,7 +57,21 @@ class ContainerViewController: UIViewController {
         })
     }
     
-
+    private func fetchStudent() {
+        if student == nil {
+            guard let currentUsersUid = Auth.auth().currentUser?.uid else {return}
+            let studentsCollectionReference = database.collection("students").document(currentUsersUid)
+            studentsCollectionReference.getDocument { (document, error) in
+                if let document = document {
+                    if let dataDescription = document.data() as? [String : String] {
+                        self.student = Student(dataDescription)
+                        self.teachersViewController.student = self.student
+                        self.sideMenuViewController.student = self.student
+                    }
+                }
+            }
+        }
+    }
 }
 
 // MARK: - TeachersViewControllerDelegate
@@ -65,6 +86,15 @@ extension ContainerViewController: SideMenuDelegate {
     func userProfileClicked() {
         self.performSegue(withIdentifier: "ShowUserProfile", sender: nil)
     }
+    
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if segue.identifier == "ShowUserProfile" {
+            let destinationVC = segue.destination.children[0] as? UserProfileViewController
+            destinationVC?.student = student
+        }
+    }
+    
 }
