+++ MusicMaker/MusicMaker/View Controllers/TeachersViewController.swift
@@ -7,8 +7,7 @@
 //
 
 import UIKit
-import FirebaseFirestore
-import Firebase
+
 
 
 class TeachersViewController: UIViewController {
@@ -18,15 +17,13 @@ class TeachersViewController: UIViewController {
         super.viewDidLoad()
         let touchGesture = UITapGestureRecognizer(target: self, action: #selector(hideMenuFromUserTap))
         self.view.addGestureRecognizer(touchGesture)
-        fetchStudent()
     }
     
     
     // MARK: - Properties
     var sideMenuIsShowing = false
-    var student: Student?
     weak var delegate: TeachersViewControllerDelegate?
-    let database = Firestore.firestore()
+    var student: Student?
     
     // MARK: - Private Methods
     @objc private func hideMenuFromUserTap() {
@@ -35,20 +32,7 @@ class TeachersViewController: UIViewController {
         }
     }
     
-    func fetchStudent() {
-        if student == nil {
-            guard let currentUsersUid = Auth.auth().currentUser?.uid else {return}
-            let studentsCollectionReference = database.collection("students").document(currentUsersUid)
-            studentsCollectionReference.getDocument { (document, error) in
-                if let document = document {
-                    if let dataDescription = document.data() as? [String : String] {
-                        self.student = Student(dataDescription)
-                        print(self.student?.firstName)
-                    }
-                }
-            }
-        }
-    }
+
     
     // MARK: - IBActions
     @IBAction func showSideMenu(_ sender: Any) {
